+++ AngularApp/projects/diagnostic-data/src/lib/models/betaSubscriptions.ts
